SQLiteLoader fixed, CDynamicsUpdater fully implemented
The CDynamicsUpdater::UpdateStates(inputs, state) is now working well. There is still no error handling for managing inputs beyong their limits (constrol surfaces beyond +/- 18 deg and motors beyond max thrust). Need to implement a way to take a file that is a reference path and build a PID controller to control the input vector for a simulation, then create a way to export the data well and fit SINDYc in matlab.